added 3 editor tests and removed old editor test that did not work
